switched to using 32 bits in client/server, and then netcode deals with only writing 16 bits where appropriate

+++ code/common.cpp
@@ -18,7 +18,7 @@ typedef double float64;
 
 constexpr uint16 	c_port 				= 9999;
 constexpr uint32 	c_socket_buffer_size = 1024;
-constexpr uint16	c_max_clients 		= 32;
+constexpr uint32	c_max_clients 		= 32;
 constexpr uint32	c_ticks_per_second 	= 60;
 constexpr float32	c_seconds_per_tick 	= 1.0f / (float32)c_ticks_per_second;
 constexpr float32 	c_turn_speed 		= 1.0f;	// how fast player turns
